Day 3
Learned a new way of data cleaning using re module and learned how to construct stacked bar chart. Also to label the stacked bar chart. Finally understood how the lambda function works. Learned to construct scatter plot with seaborn and matplot. Also, I learned how to construct the pivot table. I learned how to get the top 5 results from the table and use them to construct a figure. Learned how to choose a certain value from the list of value and construct a figure with that.